reorganize collection utilities and update imports Move slice utility functions to internal/collections/slice package Update import paths in affected files Rename 'collections' package to 'slice' for clarity
